+++ java/client/src/test/java/org/signal/libsignal/net/AuthAccountsServiceTest.kt
@@ -1,0 +1,29 @@
+//
+// Copyright 2026 Signal Messenger, LLC.
+// SPDX-License-Identifier: AGPL-3.0-only
+//
+
+package org.signal.libsignal.net
+
+import kotlinx.coroutines.test.runTest
+import org.signal.libsignal.internal.NativeTestingNice
+import kotlin.test.Test
+import kotlin.test.assertIs
+
+class AuthAccountsServiceTest {
+  @Test
+  fun testSetRegistrationLock() =
+    runTest {
+      GrpcTestCase.runTests(
+        NativeTestingNice.TESTING_SetRegistrationLockTests(),
+        AuthenticatedChatConnection::fakeConnect,
+        ::AuthAccountsService,
+        invoke = { chat, req ->
+          chat.setRegistrationLock(svrKey = SvrKey(req))
+        },
+        check = { _, actual ->
+          assertIs<RequestResult.Success<Unit>>(actual)
+        },
+      )
+    }
+}