+++ RELEASE_NOTES.md
@@ -1,2 +1,3 @@
 v0.97.4
 
+- AuthAccountsService.setRegistrationLock() typed API